+++ src/test/java/jeje/work/aeatbe/ArticleServiceTest.java
@@ -8,8 +8,8 @@ import static org.mockito.Mockito.when;
 import java.sql.Timestamp;
 import java.util.List;
 import java.util.Optional;
-import jeje.work.aeatbe.dto.column_dto.ArticleListResponseDTO;
-import jeje.work.aeatbe.dto.column_dto.ArticleResponseDTO;
+import jeje.work.aeatbe.dto.article.ArticleListResponseDTO;
+import jeje.work.aeatbe.dto.article.ArticleResponseDTO;
 import jeje.work.aeatbe.entity.Article;
 import jeje.work.aeatbe.repository.ArticleRepository;
 import jeje.work.aeatbe.service.ArticleService;
